refactor scan
