+++ src/queue/processing.ts
@@ -1,9 +1,12 @@
+import { spawn } from "child_process";
+
+import { processingCollection } from "../database/index";
+import * as logger from "../logger";
+import { getConfig } from "../config";
+
 export interface ISceneProcessingItem {
   _id: string;
 }
-
-import * as logger from "../logger";
-import { processingCollection } from "../database/index";
 
 let processing = false;
 
@@ -31,4 +34,25 @@ export async function getHead(): Promise<ISceneProcessingItem | null> {
 
 export function enqueueScene(_id: string) {
   return processingCollection.upsert(_id, { _id });
+}
+
+export async function tryStartProcessing() {
+  const config = getConfig();
+  if (!config.DO_PROCESSING) return;
+
+  const queueLen = await getLength();
+  if (queueLen > 0 && !isProcessing()) {
+    logger.message("Starting processing worker...");
+    setProcessingStatus(true);
+    spawn(process.argv[0], process.argv.slice(1).concat(["--process-queue"]), {
+      cwd: process.cwd(),
+      detached: false,
+      stdio: "inherit",
+    }).on("exit", (code) => {
+      logger.warn(`Processing process exited with code ${code}`);
+      setProcessingStatus(false);
+    });
+  } else if (!queueLen) {
+    logger.success("No more videos to process.");
+  }
 }
